Implement a indicator/control for the autoscrolling
This is intended to be used to both indicate the position of the footswitch/
controller and adjust the value that should be considered "stationary".

Also included in this is a checkable button to "ignore" the input for when a
user wants to disable the external input. This is checked by default on app
load, for what should be an obvious reason.