Simplify valid filename determination
The original code required .svg files to have a double-barrelled extension
(`.svg.p`, `.svg.pie`, or `.svg.pal`) to force Qt to use the PaletteIconEngine
instead of Qt's own SVG renderer. This was required due to how the original
code slotted into the Qt ecosystem.

As we are explicitly using the PaletteIconEngine, we do not require this.

And `os.path` can provide what we need as far as file existence goes.